Updated working with session

+++ server/api/controllers/categories/controller.ts
@@ -6,8 +6,11 @@ export class Controller {
     DbService.getAllCategories().then(r => res.json(r));
   }
 
-  create(req: Request, res: Response): void {
-    DbService.createCategory(req.body).then(r =>
+  create(req: Request, res: Response) {
+    if (!req['session'] || !req['session'].email) {
+      return res.status(401).json({ error: "401 Unauthorized" });
+    }
+    DbService.createCategory(req.body, req['session']).then(r =>
       res
         .status(201)
         .location(`/api/v1/categories/${r}`)
@@ -15,8 +18,11 @@ export class Controller {
     );
   }
 
-  update(req: Request, res: Response): void {
-    DbService.updateCategory(req.body).then(r =>
+  update(req: Request, res: Response) {
+    if (!req['session'] || !req['session'].email) {
+      return res.status(401).json({ error: "401 Unauthorized" });
+    }
+    DbService.updateCategory(req.body, req['session']).then(r =>
       res
         .status(201)
         .location(`/api/v1/categories/${r}`)
@@ -24,7 +30,10 @@ export class Controller {
     );
   }
 
-  delete(req: Request, res: Response): void {
+  delete(req: Request, res: Response) {
+    if (!req['session'] || !req['session'].email) {
+      return res.status(401).json({ error: "401 Unauthorized" });
+    }
     DbService.deleteCategory(req.body.id).then(r =>
       res
         .status(201)
